fix: put mesh_participants test where it should be
